requirements.txt
Django~=5.1.6
python-decouple~=3.8
python-decouple
django-compressor
django-libsass
python-dateutil
celery